+++ db-collections/PermissionedDataPersister.js
@@ -1,6 +1,6 @@
 "use strict";
 var Q = require("q");
-/** {@link DataPersister.Adapter} wrapper that checks permissions before reading/writing data
+/** DataPersister wrapper that checks permissions before reading/writing data
  * @author TeamworkGuy2
  */
 var PermissionedDataPersisterAdapter = (function () {
@@ -9,6 +9,16 @@ var PermissionedDataPersisterAdapter = (function () {
         this.storeSettings = storeSettings;
         this.persister = persister;
     }
+    PermissionedDataPersisterAdapter.prototype.getCollectionNames = function () {
+        if (this.syncSettings.readAllow) {
+            return this.persister.getCollectionNames();
+        }
+        else {
+            var dfd = Q.defer();
+            dfd.reject("permission denied: data persister read permissions are denied");
+            return dfd.promise;
+        }
+    };
     PermissionedDataPersisterAdapter.prototype.persist = function () {
         if (this.syncSettings.writeAllow) {
             return this.persister.persist({ compress: this.storeSettings.compressLocalStores });
@@ -26,6 +36,39 @@ var PermissionedDataPersisterAdapter = (function () {
         else {
             var dfd = Q.defer();
             dfd.reject("permission denied: data persister read permissions are denied");
+            return dfd.promise;
+        }
+    };
+    PermissionedDataPersisterAdapter.prototype.getCollectionRecords = function (collectionName, options) {
+        if (this.syncSettings.readAllow) {
+            return this.persister.getCollectionRecords(collectionName, { decompress: (options != null ? options.decompress : this.storeSettings.compressLocalStores) });
+        }
+        else {
+            var dfd = Q.defer();
+            dfd.reject("permission denied: data persister read permissions are denied");
+            return dfd.promise;
+        }
+    };
+    PermissionedDataPersisterAdapter.prototype.addCollectionRecords = function (collectionName, options, records, removeExisting) {
+        if (this.syncSettings.writeAllow) {
+            return this.persister.addCollectionRecords(collectionName, {
+                compress: (options != null ? options.compress : this.storeSettings.compressLocalStores),
+                maxObjectsPerChunk: (options != null ? options.maxObjectsPerChunk : null)
+            }, records, removeExisting);
+        }
+        else {
+            var dfd = Q.defer();
+            dfd.reject("permission denied: data persister write permissions are denied");
+            return dfd.promise;
+        }
+    };
+    PermissionedDataPersisterAdapter.prototype.clearCollections = function (collectionNames) {
+        if (this.syncSettings.writeAllow) {
+            return this.persister.clearCollections(collectionNames);
+        }
+        else {
+            var dfd = Q.defer();
+            dfd.reject("permission denied: data persister write permissions are denied");
             return dfd.promise;
         }
     };